feat: change billable_hours to int and add spin ctrl to set option

+++ res/migrations/20260829173240_add_billable_hours_to_projects_table.sql
@@ -1,1 +1,1 @@
-ALTER TABLE projects ADD COLUMN billable_hours REAL NULL;
+ALTER TABLE projects ADD COLUMN billable_hours INTEGER NULL;
